[0074] 将 s7.c 中的 continuation 相关代码迁移到 s7_continuation.c
